move:#52 공통 컴포넌트 UI 폴더로 이동 - 코드 컨벤션
